# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Creator Toolbox API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tool {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  description: string;
  shortDescription: string;
  icon: string;
  usageCount: number;
  isActive: boolean;
  createdAt: string;
}

export type ToolDetail = Tool & {
  howToGuide: string;
  seoContent: string;
  faqContent: string;
  exampleOutputs: string;
  relatedTools: Tool[];
};

export interface ToolListResponse {
  tools: Tool[];
  total: number;
}

export type ToolExecuteRequestInputs = { [key: string]: unknown };

export interface ToolExecuteRequest {
  inputs: ToolExecuteRequestInputs;
}

export type ToolExecuteResponseResult = { [key: string]: unknown };

export interface ToolExecuteResponse {
  result: ToolExecuteResponseResult;
  outputs: string[];
  success: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  icon: string;
  toolCount: number;
  color: string;
}

export type CategoryDetail = Category & {
  tools: Tool[];
  seoContent: string;
};

export interface CategoryListResponse {
  categories: Category[];
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  author: string;
  tags: string[];
  metaTitle: string;
  metaDescription: string;
  publishedAt: string;
  isPublished: boolean;
  readingTime: number;
}

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
}

export interface CreateBlogPostRequest {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  author: string;
  tags: string[];
  metaTitle: string;
  metaDescription: string;
  isPublished: boolean;
}

export interface UpdateBlogPostRequest {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  author?: string;
  tags?: string[];
  metaTitle?: string;
  metaDescription?: string;
  isPublished?: boolean;
}

export interface CreateToolRequest {
  name: string;
  slug: string;
  categoryId: number;
  description: string;
  shortDescription: string;
  icon: string;
  howToGuide?: string;
  seoContent?: string;
  faqContent?: string;
  exampleOutputs?: string;
  isActive: boolean;
}

export interface UpdateToolRequest {
  name?: string;
  slug?: string;
  categoryId?: number;
  description?: string;
  shortDescription?: string;
  icon?: string;
  howToGuide?: string;
  seoContent?: string;
  faqContent?: string;
  exampleOutputs?: string;
  isActive?: boolean;
}

export interface TrackUsageRequest {
  toolId?: number;
  toolSlug: string;
}

export type SitemapResponseUrlsItem = {
  loc: string;
  lastmod: string;
  changefreq: string;
  priority: number;
};

export interface SitemapResponse {
  urls: SitemapResponseUrlsItem[];
}

export type AdminStatsTopToolsItem = {
  name: string;
  slug: string;
  usageCount: number;
};

export interface AdminStats {
  totalTools: number;
  totalCategories: number;
  totalBlogPosts: number;
  totalUsageToday: number;
  totalUsageMonth: number;
  topTools: AdminStatsTopToolsItem[];
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListToolsParams = {
  category?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type GetPopularToolsParams = {
  limit?: number;
};

export type ListBlogPostsParams = {
  limit?: number;
  offset?: number;
  tag?: string;
};