# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Creator Toolbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get category with its tools
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category details with tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDetail"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blog post list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /analytics/track:
    post:
      operationId: trackToolUsage
      tags: [analytics]
      summary: Track tool usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackUsageRequest"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/tools:
    get:
      operationId: adminListTools
      tags: [admin]
      summary: Admin list all tools
      responses:
        "200":
          description: All tools for admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
    post:
      operationId: adminCreateTool
      tags: [admin]
      summary: Create a new tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
      responses:
        "201":
          description: Tool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
  /admin/tools/{id}:
    put:
      operationId: adminUpdateTool
      tags: [admin]
      summary: Update a tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateToolRequest"
      responses:
        "200":
          description: Tool updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
    delete:
      operationId: adminDeleteTool
      tags: [admin]
      summary: Delete a tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/blog:
    post:
      operationId: adminCreateBlogPost
      tags: [admin]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostRequest"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /admin/blog/{id}:
    put:
      operationId: adminUpdateBlogPost
      tags: [admin]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostRequest"
      responses:
        "200":
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: adminDeleteBlogPost
      tags: [admin]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    Tool:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        icon:
          type: string
        usageCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, slug, categoryId, categoryName, categorySlug, description, shortDescription, icon, usageCount, isActive, createdAt]
...
    ToolListResponse:
      type: object
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        total:
          type: integer
      required: [tools, total]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        toolCount:
          type: integer
        color:
          type: string
      required: [id, name, slug, description, icon, toolCount, color]
    CategoryDetail:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            tools:
              type: array
              items:
                $ref: "#/components/schemas/Tool"
            seoContent:
              type: string
          required: [tools, seoContent]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        publishedAt:
          type: string
        isPublished:
          type: boolean
        readingTime:
          type: integer
      required: [id, title, slug, excerpt, content, author, tags, metaTitle, metaDescription, publishedAt, isPublished, readingTime]
    BlogListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
      required: [posts, total]
    CreateBlogPostRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        isPublished:
          type: boolean
      required: [title, slug, excerpt, content, author, tags, metaTitle, metaDescription, isPublished]
    UpdateBlogPostRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        isPublished:
          type: boolean
    CreateToolRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        shortDescription:
          type: string
        icon:
          type: string
        howToGuide:
          type: string
        seoContent:
          type: string
        faqContent:
          type: string
        exampleOutputs:
          type: string
        isActive:
          type: boolean
      required: [name, slug, categoryId, description, shortDescription, icon, isActive]
    UpdateToolRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        shortDescription:
          type: string
        icon:
          type: string
        howToGuide:
          type: string
        seoContent:
          type: string
        faqContent:
          type: string
        exampleOutputs:
          type: string
        isActive:
          type: boolean
    TrackUsageRequest:
      type: object
      properties:
        toolId:
          type: integer
        toolSlug:
          type: string
      required: [toolSlug]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]