# lib/api-zod/src/generated/types/blogListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Creator Toolbox API
 * OpenAPI spec version: 1.0.0
 */
import type { BlogPost } from "./blogPost";

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
}