# lib/api-zod/src/generated/types/trackUsageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Creator Toolbox API
 * OpenAPI spec version: 1.0.0
 */

export interface TrackUsageRequest {
  toolId?: number;
  toolSlug: string;
}